requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==24.3
djangorestframework==3.15.2
Pillow==10.0.0
PyQRCode==1.2.1
reportlab==4.0.4
sqlparse==0.5.1
django-safedelete